refactor : 로그인 시 내가 설정한 동네 text 추가
