+++ meerket/meerket-application/src/main/java/org/j1p5/api/auth/dto/LoginResponse.java
@@ -1,13 +1,15 @@
 package org.j1p5.api.auth.dto;
 
+import org.j1p5.domain.activityArea.dto.SimpleAddress;
 import org.j1p5.domain.user.UserInfo;
 
-public record LoginResponse(String nickname, String profileUrl, Long activityAreaId) {
+public record LoginResponse(String nickname, String profileUrl, Integer emdId, String emdName) {
     public static LoginResponse from(UserInfo userInfo) {
         return new LoginResponse(
                 userInfo.nickname(),
                 userInfo.imageUrl(),
-                userInfo.activityAreaId()
+                userInfo.activityAreaInfo().emdId(),
+                userInfo.activityAreaInfo().emdName()
         );
     }
 }
