+++ meerket/meerket-application/src/main/java/org/j1p5/api/auth/service/OauthService.java
@@ -2,7 +2,6 @@ package org.j1p5.api.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.j1p5.domain.auth.dto.OauthProfile;
-import org.j1p5.domain.user.UserInfo;
 import org.j1p5.domain.user.entity.Provider;
 import org.j1p5.domain.user.entity.Role;
 import org.j1p5.domain.user.entity.UserEntity;
