+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/activityArea/dto/SimpleAddress.java
@@ -1,4 +1,9 @@
 package org.j1p5.domain.activityArea.dto;
 
+import org.j1p5.domain.activityArea.entity.ActivityArea;
+
 public record SimpleAddress(Integer emdId, String emdName) {
+    public static SimpleAddress from(ActivityArea activityArea) {
+        return new SimpleAddress(activityArea.getEmdArea().getId(), activityArea.getEmdArea().getEmdName());
+    }
 }
